pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mechinterp-pipeline"
version = "0.1.0"
description = "Tuned-Lens-based internal comparison pipeline"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "transformers>=4.44",
  "accelerate>=0.34",
  "tuned-lens>=0.2",
  "pandas>=2.2",
  "numpy>=1.26",
  "plotly>=5.24",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "scipy>=1.13",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
